+++ src/store/index.js
@@ -10,7 +10,6 @@ import navigation from './modules/navigation'
 import stats from './modules/stats'
 import author from './modules/author'
 import search from './modules/search'
-import darkMode from './modules/darkMode'
 
 Vue.use(Vuex)
 
@@ -26,7 +25,6 @@ export default new Vuex.Store({
     navigation,
     stats,
     author,
-    search,
-    darkMode
+    search
   }
 })