+++ src/views/Profile.vue
@@ -4,19 +4,20 @@
       <h1>{{ $t('profile') }}</h1>
     </b-container>
     <password />
-    <dark-mode />
+    <b-container size="m">
+      <h2>{{ $t('themes') }}</h2>
+      <b-theme />
+    </b-container>
   </article>
 </template>
 
 <script>
 import Password from '../components/profile/Password'
-import DarkMode from '../components/profile/DarkMode'
 
 export default {
   name: 'profile-view',
   components: {
-    Password,
-    DarkMode
+    Password
   }
 }
 </script>